refactor: renamed parser.py
